Image width and height vars changed to strong type
I used strong types in a really simple but useful way.
I made the image width and height types strong with NamedType.

The idea was that to help the developer with the NamedType, because it
is easy to mess up two integer values next to each other in a function
call.

I fixed everything related to the new typing. It includes:
- image.h
- image.cpp
- diagram.cpp

+ I added strongtype.hpp to implement the named type strong type class.